Fix Default Profile Pic Bug
[Problem]
The default profile pictures on the Oculus forums do not start with HTTP
or HTTPS.

[Solution]
Check if the URL starts with HTTP and if not, add HTTPS to the front of
it. This will still fail if there is no HTTPS version, but this seems to
never be the case on the forums, and is more secure.

[Testing]
Tested locally.