Update Error Messages
[Problem]
When a user is non-existent, an error message should be shown in
accordance with that. Similarly, errors should be more detailed for when
a Start member doesn't have their forum profile setup fully.

[Solution]
Fixed a bug with how the bot detects 404 pages for users. Also updated
the error message for users lacking the Oculus Start badge on their
forum profiles.

[Testing]
Tested locally and confirmed working.