Add Unverified Role Support
[Problem]
When a new user joins, we want to have an "Unverified" role that we can
ping to remind folks to verify occasionally before big events.

[Solution]
Setup the script to add the "Unverified" role when a new user joins. It
is removed when the user is validated.

[Testing]
Tested via the bot locally.