Add Unverify Command
[Problem]
Right now, we have a ton of existing unverified users. We need a bot
command to update all of these unverified users with the proper
Unverified role.

[Solution]
Added the unverify command to do this. Can only be run by those with the
Admin role.

[Testing]
Tested locally and updated the existing users with this.