Update Architecture Plans
[Problem]
Thanks to some newly discovered Discord APIs, it is now possible to
directly update the Discord server as soon as an event occurs, rather
than waiting for the bot to see the changes.

[Solution]
Updated the architecture of the bot to instead reach out to update
server roles and such directly, greatly speeding up our ability to
perform various actions.