Add Discord Rest API Support
[Problem]
Currently, while users can be authorized as Start members, this does not
actually do anything on the Discord server. We should use Discord's Rest
APIs to actually update the user's role accordingly.

[Solution]
Added support for Discord APIs to get a DiscordMember object from a
handle, as well as to update roles for a given DiscordMember based on
the role's human-readable name. Added support for this in the UserAuth
API, and removed the SQS queue that was there previously.

[Testing]
Added more unit tests (and updated existing ones), plus did lots of
manual testing.