Update Primary Key + Fix Secrets
[Problem]
The primary key should be the Discord user's ID, not their handle as
this can be changed. Plus there is a bug where we are creating 2 Secrets
instead!

[Solution]
Updated the primary key for the user auth table to instead be based on a
user's ID, and removed the extra Secret being created.

[Testing]
Updated unit tests + manually tested.