Add Domain Support
[Problem]
Right now, whenever the API Gateway resources are created, it could
potentially generate a new name for the resource, resulting in the
inability to contact the old resource name. This should instead be
placed behind an actual domain for consistent endpoints.

[Solution]
Added support for specifying an Alias record to a Route53 domain (with
an SSL certificate).

[Testing]
Updated unit tests + did manual testing.