Update Tools
[Problem]
Certain tools need to be run locally to update things. Specifically need
to start transferring old users into the new database.

[Solution]
Separated out the tools scripts. Still not fully finished, but it's fine
as is.

[Testing]
Verified manually that the tools are working (ish, the transfer-users
script needs a bit more work).