Limit Commands of Old Bot
[Problem]
The old bot is being slowly deprecated. Need to remove a bunch of
commands that we no longer want, as well as have the error command give
an error.

[Solution]
Removed any commands that change a user's data and added an error
message for when people try to verify.

[Testing]
Did not test (bad, I know, but we are preparing to deprecate this soon).