Switch to discord-bot-cdk-construct
[Problem]
Currently, the Discord bot endpoint is all setup within this package.

[Solution]
Switched to discord-bot-cdk-construct, a CDK construct that makes
deploying Discord bots easier. Allows this project to be simpler and
cleaner as a result.

[Testing]
Re-ran unit tests, and manually deployed the setup. Confirmed endpoints
are working as expected with it.